+++ classes/local/certification.php
@@ -48,15 +48,14 @@ final class certification {
     /**
      * Options for editing of certification descriptions.
      *
-     * @param int $contextid
      * @return array
      */
-    public static function get_description_editor_options(int $contextid): array {
+    public static function get_description_editor_options(): array {
         global $CFG;
         require_once($CFG->dirroot . '/lib/formslib.php');
 
-        $context = \context::instance_by_id($contextid);
-        return ['maxfiles' => EDITOR_UNLIMITED_FILES, 'maxbytes' => get_site()->maxbytes, 'context' => $context];
+        $syscontext = \context_system::instance();
+        return ['maxfiles' => EDITOR_UNLIMITED_FILES, 'maxbytes' => get_site()->maxbytes, 'context' => $syscontext];
     }
 
     /**
@@ -154,7 +153,7 @@ final class certification {
         }
 
         if ($editorused) {
-            $editoroptions = self::get_description_editor_options($data->contextid);
+            $editoroptions = self::get_description_editor_options();
             $data = file_postupdate_standard_editor(
                 $data,
                 'description',
@@ -242,7 +241,7 @@ final class certification {
         if (isset($data->description_editor)) {
             $data->description = $data->description_editor['text'];
             $data->descriptionformat = $data->description_editor['format'];
-            $editoroptions = self::get_description_editor_options($context->id);
+            $editoroptions = self::get_description_editor_options();
             $data = file_postupdate_standard_editor(
                 $data,
                 'description',
@@ -313,21 +312,6 @@ final class certification {
 
         $trans = $DB->start_delegated_transaction();
 
-        get_file_storage()->move_area_files_to_new_context(
-            $certification->contextid,
-            $context->id,
-            'tool_mucertify',
-            'image',
-            $certification->id
-        );
-        get_file_storage()->move_area_files_to_new_context(
-            $certification->contextid,
-            $context->id,
-            'tool_mucertify',
-            'description',
-            $certification->id
-        );
-
         // Do not check if tags enabled here.
         $tags = \core_tag_tag::get_item_tags_array('tool_mucertify', 'tool_mucertify_certification', $certification->id);
         if ($tags) {
@@ -360,11 +344,11 @@ final class certification {
         global $DB;
 
         $certification = $DB->get_record('tool_mucertify_certification', ['id' => $data->id], '*', MUST_EXIST);
-        $context = \context::instance_by_id($certification->contextid);
+        $syscontext = \context_system::instance();
 
         if (isset($data->image)) {
-            file_save_draft_area_files($data->image, $context->id, 'tool_mucertify', 'image', $data->id, ['subdirs' => 0, 'maxfiles' => 1]);
-            $files = get_file_storage()->get_area_files($context->id, 'tool_mucertify', 'image', $data->id, '', false);
+            file_save_draft_area_files($data->image, $syscontext->id, 'tool_mucertify', 'image', $data->id, ['subdirs' => 0, 'maxfiles' => 1]);
+            $files = get_file_storage()->get_area_files($syscontext->id, 'tool_mucertify', 'image', $data->id, '', false);
             $presenation = (array)json_decode($certification->presentationjson);
             if ($files) {
                 $file = reset($files);
@@ -389,12 +373,12 @@ final class certification {
     public static function get_image_url(stdClass $certification, bool $generateifmissing): ?url {
         global $CFG;
 
+        $syscontext = \context_system::instance();
         $presentation = (array)json_decode($certification->presentationjson);
         if (!empty($presentation['image'])) {
-            $context = \context::instance_by_id($certification->contextid);
             return url::make_file_url(
                 "$CFG->wwwroot/pluginfile.php",
-                '/' . $context->id . '/tool_mucertify/image/' . $certification->id . '/' . $presentation['image']
+                '/' . $syscontext->id . '/tool_mucertify/image/' . $certification->id . '/' . $presentation['image']
             );
         }
 
@@ -402,7 +386,6 @@ final class certification {
             return null;
         }
 
-        $syscontext = \context_system::instance();
         return url::make_file_url(
             "$CFG->wwwroot/pluginfile.php",
             '/' . $syscontext->id . '/tool_mucertify/image/' . $certification->id . '/geopattern.svg'
@@ -774,7 +757,7 @@ final class certification {
         $trans = $DB->start_delegated_transaction();
 
         $certification = $DB->get_record('tool_mucertify_certification', ['id' => $id], '*', MUST_EXIST);
-        $context = \context::instance_by_id($certification->contextid);
+        $syscontext = \context_system::instance();
 
         // Delete notifications configuration and data.
         notification_manager::delete_certification_notifications($certification);
@@ -791,10 +774,11 @@ final class certification {
         $DB->delete_records('tool_mucertify_period', ['certificationid' => $certification->id]);
 
         // Certification details last.
-        \core_tag_tag::set_item_tags('tool_mucertify', 'tool_mucertify_certification', $certification->id, $context, null);
+        \core_tag_tag::remove_all_item_tags('tool_mucertify', 'tool_mucertify_certification', $certification->id);
+
         $fs = get_file_storage();
-        $fs->delete_area_files($context->id, 'tool_mucertify', 'description', $certification->id);
-        $fs->delete_area_files($context->id, 'tool_mucertify', 'image', $certification->id);
+        $fs->delete_area_files($syscontext->id, 'tool_mucertify', 'description', $certification->id);
+        $fs->delete_area_files($syscontext->id, 'tool_mucertify', 'image', $certification->id);
 
         $DB->delete_records('tool_mucertify_certification', ['id' => $certification->id]);